12 - Schema Format Change and Resolver Restructuring
This commit changes the schema as described here:
https://github.com/trynmaps/tryn-api/issues/12
along with removing some currently unused types.

Moreover, it simplifies the resolver by grouping
the vehicles by rid and then by vtime in one go.

`makeStopsFromNextBus()` was moved from formatStateData
to getStops - and formatStateData was removed as it
is no longer needed.

More commenting was also added as to make this
part of the API (among the most convoluted section)
easier to read and understand.

Tests:

There are currently no automated tests for our API;
however, we can manually test it by running the
query posted in the issue link above.