41 - Add muni metro duplicate removal (#42)
Make use of leadingVid field in vehicle
by removing 'follower' vehicles from the
map, so that only one car of a 2-car train
is returned. A numCars field is also returned
and added to the schema, which has the number
of cars (one car if the field is undefined).

For all vehicles being processed in a state,
if A leads B in one state, A will lead B in
all states given its vid/rid/did stays the same.
If a 'follower' (vehicle with a leadingVid) is in
a state but its leader is not, then it is renamed.
There are also checks to address cycles and chaining
of the leadingVid field.

Lastly, vehicles from orion-raw have vid/rid/did/leadingVid
fields added to them as to use for testing.